CHANGE: Move Touch API into it's own file
