+++ Packages/com.unity.input/Runtime/Input_Touch.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 472bf9b836084ab49ba8ecbc22229965
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
